requirements.txt
deeppavlov
stanfordcorenlp
py2neo
psycopg2-binary
beautifulsoup4
pyjnius
nltk